+++ src/splat/segtypes/psx/segment.py
@@ -1,5 +1,0 @@
-from ..segment import Segment
-
-
-class PsxSegment(Segment):
-    pass