+++ src/splat/segtypes/segment.py
@@ -446,8 +446,7 @@ class Segment:
 
     @property
     def subalign(self) -> Optional[int]:
-        if self.parent:
-            log.error("subalign is not valid for non-top-level segments")
+        assert self.parent is None, f"subalign is not valid for non-top-level segments. ({self})"
         return self.given_subalign
 
     @property