+++ src/splat/segtypes/n64/__init__.py
@@ -14,12 +14,10 @@ from . import ia4 as ia4
 from . import ia8 as ia8
 from . import img as img
 from . import ipl3 as ipl3
-from . import linker_offset as linker_offset
 from . import mio0 as mio0
 from . import palette as palette
 from . import rgba16 as rgba16
 from . import rgba32 as rgba32
 from . import rsp as rsp
-from . import segment as segment
 from . import vtx as vtx
 from . import yay0 as yay0