+++ src/splat/segtypes/n64/img.py
@@ -4,10 +4,10 @@ from typing import Dict, List, Tuple, Type, Optional, Union
 from n64img.image import Image
 from ...util import log, options
 
-from .segment import N64Segment
+from ..segment import Segment
 
 
-class N64SegImg(N64Segment):
+class N64SegImg(Segment):
     @staticmethod
     def parse_dimensions(yaml: Union[Dict, List]) -> Tuple[int, int]:
         if isinstance(yaml, dict):
@@ -88,7 +88,7 @@ class N64SegImg(N64Segment):
     @staticmethod
     def estimate_size(yaml: Union[Dict, List]) -> int:
         width, height = N64SegImg.parse_dimensions(yaml)
-        typ = N64Segment.parse_segment_type(yaml)
+        typ = Segment.parse_segment_type(yaml)
 
         if typ == "ci4" or typ == "i4" or typ == "ia4":
             return width * height // 2