+++ src/splat/segtypes/n64/ipl3.py
@@ -1,4 +1,3 @@
-from ..common.code import CommonSegCode
 from ..common.hasm import CommonSegHasm