+++ src/splat/segtypes/psp/__init__.py
@@ -1,1 +1,0 @@
-from . import segment as segment