+++ src/splat/segtypes/psp/segment.py
@@ -1,5 +1,0 @@
-from ..segment import Segment
-
-
-class PspSegment(Segment):
-    pass