+++ src/splat/segtypes/psx/__init__.py
@@ -1,3 +1,2 @@
 from . import asm as asm
 from . import header as header
-from . import segment as segment