+++ src/splat/segtypes/segment.py
@@ -298,8 +298,8 @@ class Segment:
             yaml
         )
 
-        # True if this segment was generated based on auto_all_sections
-        self.is_auto_all: bool = False
+        # True if this segment was generated based on auto_link_sections
+        self.is_generated: bool = False
 
         if self.rom_start is not None and self.rom_end is not None:
             if self.rom_start > self.rom_end:
@@ -446,10 +446,10 @@ class Segment:
 
     @property
     def subalign(self) -> Optional[int]:
-        if self.parent:
-            return self.parent.subalign
-        else:
-            return self.given_subalign
+        assert (
+            self.parent is None
+        ), f"subalign is not valid for non-top-level segments. ({self})"
+        return self.given_subalign
 
     @property
     def vram_symbol(self) -> Optional[str]: