+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### 0.27.0
 
 * BREAKING: Renamed `auto_all_sections` to `auto_link_sections` and documented its behavior.
-* BREAKING: Removed redundant `N64Segment` stub class. Any references to this should be instead to the base `Segment`
+* BREAKING: Removed redundant `N64Segment`, `PSXSegment`, `PSPSegment` stub classes. Any references to these should be instead to the base `Segment`
 * Promoted `linker_offset` segment type to common, so it's now usable by all platforms.
 * Added documentation for the remaining undocumented segment types and did some general doc tidying.
 * Splat will now error when the last segment is `pad`, as this will not work as expected.